# openapi/canonical/sources.json
{"components":{"headers":{"BearerChallenge":{"description":"Authentication scheme required by the Server.","schema":{"example":"Bearer","type":"string"}},"Location":{"description":"URI of the newly created Source or Artifact head.","schema":{"type":"string"}},"RequestId":{"description":"Opaque identifier for correlating one request.","schema":{"type":"string"}}},"responses":{"Conflict":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"The command conflicts with current immutable state.","headers":{"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"InternalError":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"The Server failed without exposing internal details.","headers":{"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"InvalidRequest":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"The request violates the transport or application contract.","headers":{"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"NotFound":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"The requested durable value was not found or is not observable.","headers":{"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"Unauthorized":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"A valid bearer token is required by this Server deployment.","headers":{"WWW-Authenticate":{"$ref":"#/components/headers/BearerChallenge"},"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"Unavailable":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"A required Runtime binding or dependency is unavailable.","headers":{"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}}},"schemas":{"CreateSourceRequest":{"additionalProperties":false,"properties":{"content":{"description":"JSON value persisted by the built-in content Source adapter."},"source_type":{"default":"content","enum":["content"],"type":"string"}},"required":["content"],"type":"object"},"ErrorDetail":{"additionalProperties":false,"properties":{"code":{"type":"string"},"details":{"additionalProperties":true,"nullable":true,"type":"object"},"message":{"type":"string"}},"required":["code","message","details"],"type":"object"},"ErrorResponse":{"additionalProperties":false,"properties":{"error":{"$ref":"#/components/schemas/ErrorDetail"}},"required":["error"],"type":"object"},"SourceRecord":{"properties":{"content":{"description":"Persisted canonical JSON content."},"content_digest":{"pattern":"^sha256:[0-9a-f]{64}$","type":"string"},"position":{"minimum":1,"type":"integer"},"scope_id":{"maxLength":256,"minLength":1,"pattern":".*\\S.*","type":"string"},"source_id":{"maxLength":256,"minLength":1,"pattern":"^[\\x21-\\x7E]+$","type":"string"},"source_type":{"enum":["content"],"type":"string"}},"required":["scope_id","source_type","source_id","content","position","content_digest"],"type":"object"}},"securitySchemes":{"BearerAuth":{"description":"Static bearer token used when local Server authentication is enabled.","scheme":"bearer","type":"http"}}},"info":{"description":"Remote PowerContext transport. Runtime behavior is reported by /v1/capabilities.","title":"PowerContext API","version":"0.1.0"},"openapi":"3.0.3","paths":{"/v1/scopes/{scope_id}/sources":{"post":{"description":"Persist one Source without synchronously deriving Artifacts. The Server generates source_id.","operationId":"create_source","parameters":[{"in":"path","name":"scope_id","required":true,"schema":{"maxLength":256,"minLength":1,"pattern":".*\\S.*","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSourceRequest"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceRecord"}}},"description":"The Source was durably created.","headers":{"Location":{"$ref":"#/components/headers/Location"},"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"401":{"$ref":"#/components/responses/Unauthorized"},"409":{"$ref":"#/components/responses/Conflict"},"422":{"$ref":"#/components/responses/InvalidRequest"},"500":{"$ref":"#/components/responses/InternalError"},"503":{"$ref":"#/components/responses/Unavailable"}},"summary":"Create a durable Source","tags":["sources"]}},"/v1/scopes/{scope_id}/sources/{source_type}/{source_id}":{"get":{"operationId":"get_source","parameters":[{"in":"path","name":"scope_id","required":true,"schema":{"maxLength":256,"minLength":1,"pattern":".*\\S.*","type":"string"}},{"in":"path","name":"source_type","required":true,"schema":{"enum":["content"],"type":"string"}},{"in":"path","name":"source_id","required":true,"schema":{"maxLength":256,"minLength":1,"pattern":"^[\\x21-\\x7E]+$","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceRecord"}}},"description":"The exact Source.","headers":{"X-PowerContext-Request-ID":{"$ref":"#/components/headers/RequestId"}}},"401":{"$ref":"#/components/responses/Unauthorized"},"404":{"$ref":"#/components/responses/NotFound"},"422":{"$ref":"#/components/responses/InvalidRequest"},"500":{"$ref":"#/components/responses/InternalError"},"503":{"$ref":"#/components/responses/Unavailable"}},"summary":"Get one exact Source","tags":["sources"]}}},"security":[{"BearerAuth":[]},{}]}